Commit changes

+++ GlobalTicket.TicketManagement.Application/Features/Events/Commands/DeleteEvent/DeleteEventCommandHandler.cs
@@ -1,11 +1,13 @@
-﻿using AutoMapper;
-using GlobalTicket.TicketManagement.Application.Contracts.Persistence;
-using MediatR;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using AutoMapper;
+using GlobalTicket.TicketManagement.Application.Contracts.Persistence;
+using GlobalTicket.TicketManagement.Application.Exceptions;
+using GlobalTicket.TicketManagement.Domain.Entities;
+using MediatR;
 
 namespace GlobalTicket.TicketManagement.Application.Features.Events.Commands.DeleteEvent
 {
@@ -23,6 +25,11 @@ namespace GlobalTicket.TicketManagement.Application.Features.Events.Commands.Del
         public async Task<Unit> Handle(DeleteEventCommnad request, CancellationToken cancellationToken)
         {
             var eventToDelete = await _eventRepository.GetByIdAsync(request.EventId);
+
+            if (eventToDelete == null)
+            {
+                throw new NotFoundException(nameof(Event), request.EventId);
+            }
 
             await _eventRepository.DeleteAsync(eventToDelete);
